CXP-1113: move connection persistence services
