+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/wrong_credentials_connection.yml
@@ -1,28 +1,13 @@
 services:
-    akeneo_connectivity.connection.persistence.query.are_credentials_valid_combination:
-        class: Akeneo\Connectivity\Connection\Infrastructure\Persistence\Dbal\Query\DbalAreCredentialsValidCombinationQuery
-        arguments:
-            - '@database_connection'
-
-    akeneo_connectivity.connection.persistence.query.select_connection_code_by_client_id:
-        class: Akeneo\Connectivity\Connection\Infrastructure\Persistence\Dbal\Query\DbalSelectConnectionCodeByClientIdQuery
-        arguments:
-            - '@database_connection'
-
-    akeneo_connectivity.connection.persistence.repository.wrong_credentials_combination:
-        class: Akeneo\Connectivity\Connection\Infrastructure\Persistence\Dbal\Repository\DbalWrongCredentialsCombinationRepository
-        arguments:
-            - '@database_connection'
-
     akeneo_connectivity.connection.api_connection_subscriber:
         class: Akeneo\Connectivity\Connection\Infrastructure\EventSubscriber\ApiAuthenticationEventSubscriber
         arguments:
             - '@akeneo_connectivity.connection.service.connection_context'
-            - '@akeneo_connectivity.connection.persistence.repository.wrong_credentials_combination'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Connections\WrongCredentialsCombination\Persistence\DbalWrongCredentialsCombinationRepository'
         tags:
             - { name: kernel.event_subscriber }
 
     Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\WrongCredentialsCombination\ListWrongCredentialsCombinationsAction:
         public: true
         arguments:
-            - '@akeneo_connectivity.connection.persistence.repository.wrong_credentials_combination'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Connections\WrongCredentialsCombination\Persistence\DbalWrongCredentialsCombinationRepository'
