+++ src/Akeneo/Connectivity/Connection/back/tests/Unit/spec/Infrastructure/Connections/WrongCredentialsCombination/EventSubscriber/ApiAuthenticationEventSubscriberSpec.php
